requirements.txt
pymongo[srv]>=4.6.3
python-dotenv>=1.0.1
discord.py[voice]>=2.3.2
pandas>=2.2.2
matplotlib>=3.8.4
numpy>=1.26.4
python-dateutil>=2.9.0
pytz>=2024.1
tzdata>=2024.1
youtube_dl>=2021.12.17

# pip install --upgrade --force-reinstall "git+https://github.com/ytdl-org/youtube-dl.git"